Protect forms that are autofilled by the password manager
On https://yeswehack.com/auth/login the password manager *almost* works.
It successfully autofills my username and password, but then the website
deletes the autofilled values right away. Drat. I suspect this is
probably intentional because the login form uses autocomplete="off" to
request that we not autocomplete it. Epiphany does not respect this
request because that would encourage users to use an insecure memorable
password rather than a secure password.

Unfortunately it's not currently possible to receive a notification when
the value of the form is changed by JavaScript, so my solution here is
to use a heavy hammer and just check the value of the form every 50
milliseconds for 5 seconds after the form is created. This is a safe
value that should be much longer than is actually necessary.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1865>