location-entry: bring back the custom context menu
Epiphany 49 unfortunately uses a standard GtkEntry context menu, which
is unfortunately not very good. The only customization is to add a Paste
and Go menu item to the very bottom, which is plainly the wrong place as
it's not next to the normal Paste menu item.

This brings back the good context menu used prior to Epiphany 49.

Fixes #2701