window-commands: don't take snapshot when saving PNG image
We have an easter egg that allows users to save a screenshot of the web
page by changing the file suffix to ".png" on the save dialog. This is
fun, but we don't want to do that when the main resource is a PNG image.
In that case, we surely want to save the actual PNG image.


(cherry picked from commit 03ef91d5a0a1a2cce3c4839f6145cae47574f71b)

Co-authored-by: Michael Catanzaro <mcatanzaro@redhat.com>