window: fix crash in url_should_open_automatically()
Let's not crash if the URL cannot be converted to a security origin.
This shouldn't normally happen and I'm not sure how exactly it happened
to me today, but it's not worth crashing over.

Also, rename some variables to be more specific.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1909>