Always show Quit menu item in web app mode
Multiple problems here:

The Quit action needs to always be enabled for Ctrl+Q to work, so trying
to tie the menu item visibility to whether the action is enabled was a
mistake.

We are using g_action_group_action_enabled_changed() to emit a signal
indicating that the Quit action has been disabled, but without actually
disabling the action. This is enough to trick Blueprint into thinking the
action has been enabled/disabled, even though it's always enabled. But
this is too hacky.

Instead of trying to fix this, let's just always show the menu item when
in web app mode, and not show it outside web app mode.
Conditionally-visible menu items is a source of confusion for users
anyway.

Fixes #2688

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1913>