bookmark-properties: fix on_tags_activated callback
The callback was taking wrong arguments, as the callback is supposed to
take two arguments, not three, and the user_data is in the second one.
The third one was NULL presumably because the register was zeroed.

This fixes crash when editing bookmark tags.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1915>