location-entry: use ephy-site-button-symbolic for secure websites
This icon better conveys that there is a menu here. Nobody is going to
expect to find a menu under the lock icon.

Exception: if the page is insecure, then we show the insecure lock icon
instead. Hopefully users are able to remember that this is where the
menu is.

I have a feeling this may be confusing, but it's better than *always*
showing a lock icon.