Fix duplicate bookmark entries when importing
When importing bookmarks with tags, it was possible for multiple entries of a bookmark to be listed. This is because the order updating logic was looking up existing bookmarks in an unsorted sequence. According to the docs, g_sequence_lookup() will fail if the data is unsorted.

Closes #2731

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1933>