Explicit set position of suggestions popover
In narrow mode suggestions popover tends to open to the bottom which is
not visible on mobile devices. As the url entry is now on the bottom
ensure that suggestions popover is visible by setting position to top in
this case.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1939>