Handle edge case when undoing bookmark removal from window
The bookmark removed toast that appears in the window was missing a
check. This check prevents adding the bookmark back if the user adds it
back manually before clicking undo. This check is already in the
bookmarks dialog equivalent of this toast.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1950>