shell: fix leak in ephy_shell_before_emit()
g_variant_iter_loop() is weird: normally it frees the memory for us, but
if we break out of the loop then we have to free it manually. This is
risky and we got it wrong here.

g_variant_iter_next() is simpler: there we just always have to free it
ourselves. I tried changing to use this instead, but the code was
actually a little more complicated, so wound up sticking with
g_variant_iter_loop().