Fix inconsistent cursor position when using Ctrl+K
If the entry was unfocused and Ctrl+K was pressed, the cursor would be
before the search engine bang. Subsequent uses of Ctrl+K or using it
while the entry is focused would place the cursor after. This ensures
it's placed after in all scenarios.

Closes #2732

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1972>