location-entry: consistently decode non-Latin URLs
Currently, on the page https://כולנו.ישראל/ update_url_button_style()
just fails and we wind up displaying the full URL. Moreover, if you
press Ctrl+L and then Escape to focus and unfocus the location entry,
the URL decodes itself and we wind up viewing the punycode-encoded URL
https://xn--9dbaqfu.xn--4dbrk0ce/

Fix this by replacing use of GUri with a handrolled parser that is just
good enough to get at the URL's hostname component, *without* encoding
anything. This way the entire function operates only on the decoded URL,
which is the right thing to do because we are just trying to format it
for display.

The URL decoding itself is caused because this function is trying to use
the actual text of the underlying entry, which is pointless. When the
location entry is unfocused, we always want to display the hostname of
whatever page is currently loaded. There's no point in looking at the
contents of the entry at all. We can just get it from
EphyLocationController.

This commit also introduces a helper to emit the get-location signal. I
don't like this signal, but I'm not going to attempt to remove it now. I
also revert it back to its previous behavior prior to !1980, such that
it returns the encoded URL and expects EphyLocationEntry to decode it.

Fixes #2770

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1988>