location-entry: fix file URLs disappearing
If there is no host component to the URL, we should show the full URL.
Previously this was achieved by failing and then bailing to the bottom
of the function. After bb1fe8e1470d36dde4af2829ebf5330186e1b22c, the
code improperly succeeds and continues to show only the hostname, which
is empty.

Fixes #2787

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1997>