sqlite: Start caching history service statements
At the moment, we were essentially just recompiling the SQL statement
each time we perform these functions. This is not the most efficient
way to do this. Instead, it's preferred to compile the SQL query once,
cache and use the statement, then on the next usage, it will instead
just reset the statement for usage again.

Signed-off-by: Swagtoy <me@ow.swag.toys>
Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/2009>